% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_palette_size.R
\name{adjust_palette_size}
\alias{adjust_palette_size}
\title{adjust_palette_size}
\usage{
adjust_palette_size(
  object_length,
  basic_color_palette = c("#cb4bbe", "lightskyblue", "grey37", "#53ba48", "moccasin",
    "#dcca44", "#502d70", "#afd671", "#cb4471", "#69da99", "#d44732", "#6fd3cf",
    "#5d2539", "#cfd0a0", "blue", "#d2883b", "#6793c0", "#898938", "#c98ac2", "yellow",
    "#c4c0cc", "#7d3d23", "#00a5ff", "#d68d7a", "#a2c1a3")
)
}
\arguments{
\item{object_length}{Num of parameters to color}

\item{basic_color_palette}{vector of colors to start with}
}
\value{
list of HEX colors
}
\description{
adjust_palette_size
}
