% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_population_expression_sets.R
\name{bulk_population_expression_sets}
\alias{bulk_population_expression_sets}
\title{bulk_population_expression_sets}
\usage{
bulk_population_expression_sets(bulk.contrasts)
}
\arguments{
\item{bulk.contrasts}{set of contrasts to make}
}
\value{

}
\description{
generate expression lists for gene set enrichment analysis
}
