% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multi_object_list.R
\name{create_multi_object_list}
\alias{create_multi_object_list}
\title{create_multi_object_list}
\usage{
create_multi_object_list(x)
}
\arguments{
\item{x}{directory containing barcode, gene, and matrix files along the path /outs/filtered_feature_bc_matrix/}
}
\value{

}
\description{
create_multi_object_list
}
