% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_percentMito_column.R
\name{create_percentMito_column}
\alias{create_percentMito_column}
\title{create_percentMito_column}
\usage{
create_percentMito_column(my_object)
}
\arguments{
\item{my_object}{object of Seurat class <=v2.4.3}
}
\value{

}
\description{
create_percentMito_column
}
