% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_plot_param.R
\name{cycle_plot_param}
\alias{cycle_plot_param}
\title{cycle_plot_param}
\usage{
cycle_plot_param(plotting_function, cycle_parameter, the_object)
}
\arguments{
\item{plotting_function}{oneof type "cluster", "trajectory", "qplot"}

\item{cycle_parameter}{the number of colors needed}

\item{the_object}{monocle object}
}
\value{

}
\description{
cycle_plot_param
}
