% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawmyplot.R
\name{drawmyplot}
\alias{drawmyplot}
\title{drawmyplot}
\usage{
drawmyplot(geneList, tsne.obj, name)
}
\arguments{
\item{geneList}{List of genes on which to base violin and feature plots}

\item{tsne.obj}{Seurat2 object}

\item{name}{FilePrefex}
}
\value{

}
\description{
drawmyplot
}
