% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emphasis_plots_PCA.R
\name{emphasis_plots_PCA}
\alias{emphasis_plots_PCA}
\title{emphasis_plots_PCA}
\usage{
emphasis_plots_PCA(
  object,
  file_suffix,
  iterationSlot,
  background_color = "black",
  emphasis_color = "red",
  ...
)
}
\arguments{
\item{object}{seurat object}

\item{file_suffix}{name to add to end of file}

\item{background_color}{color to give non-emphasized pops}

\item{emphasis_color}{color to give emphasized pops}

\item{...}{parameters passed on to png_plotFunction}
}
\value{

}
\description{
emphasis_plots_PCA
}
