% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoEnrichmentAnalysis.R
\name{GO_enrichment_function}
\alias{GO_enrichment_function}
\title{GO_enrichment_function}
\usage{
GO_enrichment_function(
  my_folder,
  number_of_terms = Inf,
  list.species = "Mm",
  ontology.list = c("BP", "MF", "CC"),
  file.pattern = ".txt$",
  outfile.suffix = "-GOAll.txt",
  id.type = "ensembl"
)
}
\arguments{
\item{my_folder}{folder containing files with gene lists in ensembl or symbol format}

\item{number_of_terms}{how many terms to return (use 'Inf' to return everything)}

\item{list.species}{currently only mm10 is supported}

\item{ontology.list}{ontology types to return (bioprocess, molecular function, or cell compartment)}

\item{file.pattern}{grep suffix pattern describing files in folder that contain gene lists}

\item{outfile.suffix}{suffix to add to base file name}

\item{id.type}{gene ID type (either ensmembl or symbol)}
}
\value{

}
\description{
GO_enrichment_function
}
