% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_ranking.R
\name{gsea_ranking}
\alias{gsea_ranking}
\title{gsea_ranking}
\usage{
gsea_ranking(lrt, outname, FDRmax = 0.05)
}
\arguments{
\item{lrt}{edgeR lrt object}

\item{outname}{prefix for generated files}

\item{FDRmax}{max FDR to allow}
}
\value{
file withcolumn HsHmlg identifier and logFC value at or below FDR threshold
}
\description{
generate ranked subpopulation gene lists
}
