% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSmearPlot.R
\name{makeSmearPlot}
\alias{makeSmearPlot}
\title{makeSmearPlot}
\usage{
makeSmearPlot(lrt, nameA = "nameA", nameB = "CMP", fc = FC, outname = outname)
}
\arguments{
\item{lrt}{dge object}

\item{nameA}{comparison string to describe A}

\item{nameB}{comparison string to describe  B}

\item{fc}{fold-change threshold}
}
\value{

}
\description{
makeSmearPlot
}
