% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mouse2Human_idconversion.R
\name{Mouse2Human}
\alias{Mouse2Human}
\title{Mouse2Human}
\usage{
Mouse2Human(MouseGenes)
}
\arguments{
\item{MouseGenes}{list of mouse genes}
}
\value{
table of gene identifiers
}
\description{
Credits mainly to Dave Tang's blog (https://www.biostars.org/p/147484/)
}
