% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouseIDTable}
\alias{mouseIDTable}
\title{mouseIDTable}
\format{
Table with 24117 obs. of 5 variables
\describe{
\item{Mouse.Gene_ID}{Ensemble mosue gene id}
\item{MGI}{MGI ID}
\item{Entrez.Gene_ID}{Entrez gene ID}
\item{Human.Gene_ID}{Human homologue}
\item{HGNC}{HGNC ID}
}
}
\source{
biomaRt
}
\usage{
mouseIDTable
}
\description{
A list of mouse genes and ID conversions
}
\concept{data}
\keyword{datasets}
