% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png_plotFunction.R
\name{png_plotFunction}
\alias{png_plotFunction}
\title{png_plotFunction}
\usage{
png_plotFunction(plot_to_make = x, filename = y, height = 800, width = 800)
}
\arguments{
\item{plot_to_make}{plot function}

\item{filename}{filename to for saved plot}

\item{height}{png height (px)}

\item{width}{png width (px)}
}
\value{

}
\description{
png_plotFunction
}
