% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_10XGenomics_data.R
\name{read_10XGenomics_data}
\alias{read_10XGenomics_data}
\title{read_10XGenomics_data}
\usage{
read_10XGenomics_data(x)
}
\arguments{
\item{x}{directory containing gene, barcode, and matrix files along the path /outs/filtered_gene_bc_matrices/<genome>}
}
\value{

}
\description{
read_10XGenomics_data
}
