% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seurat_args.R
\name{read_seurat_args}
\alias{read_seurat_args}
\title{read_seurat_args}
\usage{
read_seurat_args(
  data,
  projectName,
  genome,
  max_pcs,
  resolutionList,
  vars_to_regress
)
}
\value{

}
\description{
read_seurat_args
}
