% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEdgeRAnalysis.R
\name{runEdgeRAnalysis}
\alias{runEdgeRAnalysis}
\title{runEdgeRAnalysis}
\usage{
runEdgeRAnalysis(file_list, dgelist_groups, ref_sample)
}
\arguments{
\item{file_list}{Directory containing RSEM files in the format}

\item{dgelist_groups}{Character vector of sample names (for factor grouping)}

\item{ref_sample}{Element from dgelist_groups to define as reference sample}
}
\value{

}
\description{
runEdgeRAnalysis
}
