% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMonocleAnalysis.R
\name{runMonocleAnalysis}
\alias{runMonocleAnalysis}
\title{runMonocleAnalysis}
\usage{
runMonocleAnalysis(args)
}
\arguments{
\item{args}{list of arguments passed to runMonocleAnalysis function (seurat_object_filename, num_dim, cca_variables)}
}
\value{

}
\description{
runMonocleAnalysis
}
