% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSeuratAnalysis.R
\name{runSeuratAnalysis}
\alias{runSeuratAnalysis}
\title{runSeuratAnalysis}
\usage{
runSeuratAnalysis(args)
}
\arguments{
\item{args}{}
}
\value{

}
\description{
runSeuratAnalysis
}
\examples{
\dontrun{
args <- c('--data CMP-m1 CMP-m2 --projectName CMP12k_vs_CMP3k --genome mm10 --max_pcs 25 --resolutionList 0.6 --perform_cca FALSE --vars_to_regress nUMI nGene orig.ident')
runSeuratAnalysis(args)
}
}
