% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_palette_size.R
\name{adjust_palette_size}
\alias{adjust_palette_size}
\title{adjust_palette_size}
\usage{
adjust_palette_size(object_length, basic_color_palette)
}
\arguments{
\item{basic_color_palette}{}
}
\value{

}
\description{
adjust_palette_size
}
