% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_10XGenomics_data.R
\name{read_10XGenomics_data}
\alias{read_10XGenomics_data}
\title{read_10XGenomics_data}
\usage{
read_10XGenomics_data(sample.list)
}
\arguments{
\item{sample.list}{list of sample names that contain gene, barcode, and matrix files the directory}
}
\value{
named character of sample names and directory paths
}
\description{
read_10XGenomics_data
}
