% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEdgeRAnalysis.R
\name{runEdgeRAnalysis}
\alias{runEdgeRAnalysis}
\title{runEdgeRAnalysis}
\usage{
runEdgeRAnalysis(file_list, dgelist_groups, ref_sample)
}
\arguments{
\item{ref_sample}{}
}
\value{

}
\description{
Based on Stanford Tutorial at https://web.stanford.edu/class/bios221/labs/rnaseq/lab_4_rnaseq.html
}
\examples{
\dontrun{
file_list <- c(list.files(pattern = '_R', ignore.case = FALSE, recursive = FALSE, include.dirs = FALSE), list.files(pattern = '^CMP', recursive = FALSE, include.dirs = FALSE))
runEdgeRAnalysis(file_list)
}



}
