% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMonocleAnalysis.R
\name{runMonocleAnalysis}
\alias{runMonocleAnalysis}
\title{runMonocleAnalysis}
\usage{
runMonocleAnalysis(args)
}
\arguments{
\item{args}{list of arguments passed to runMonocleAnalysis function}
}
\value{

}
\description{
runMonocleAnalysis
}
\examples{
\dontrun{
args <- c('--seurat_object_filename runSeuratAnalysis.rds --num_dim 20 --cca_variables "~nUMI + nGene" "')
runMonocleAnalysis(args)
}
}
