% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSeuratAnalysis.R
\name{runSeuratAnalysis}
\alias{runSeuratAnalysis}
\title{runSeuratAnalysis}
\usage{
runSeuratAnalysis(args)
}
\arguments{
\item{args}{list of arguments (data, project name, genome, max_pcs, resolutionList, perform_cca, vars_to_regress)}
}
\value{

}
\description{
runSeuratAnalysis
}
